reorg modules: artifact/ => frameworks/
