+++ bentoml/frameworks/keras.py
@@ -1,23 +1,9 @@
-# Copyright 2019 Atalaya Tech, Inc.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-# http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 import os
 import importlib
 
-from bentoml.service_env import BentoServiceEnv
+from bentoml.service.env import BentoServiceEnv
 from bentoml.utils import cloudpickle
-from bentoml.artifact import BentoServiceArtifact
+from bentoml.service.artifacts import BentoServiceArtifact
 from bentoml.exceptions import (
     ArtifactLoadingException,
     MissingDependencyException,
@@ -56,8 +42,9 @@ class KerasModelArtifact(BentoServiceArtifact):
     >>> model_to_save.fit(...)
     >>>
     >>> import bentoml
+    >>> from bentoml.frameworks.keras import KerasModelArtifact
     >>>
-    >>> @bentoml.env(pip_dependencies=['tensorflow==1.14.0', 'numpy', 'pandas'])
+    >>> @bentoml.env(auto_pip_dependencies=True)
     >>> @bentoml.artifacts([KerasModelArtifact('model')])
     >>> class KerasModelService(bentoml.BentoService):
     >>>     @bentoml.api(input=JsonInput())
