+++ bentoml/frameworks/onnx.py
@@ -1,29 +1,15 @@
-# Copyright 2019 Atalaya Tech, Inc.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-# http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 import logging
 import os
 import pathlib
 import shutil
 
-from bentoml.artifact import BentoServiceArtifact
+from bentoml.service.artifacts import BentoServiceArtifact
 from bentoml.exceptions import (
     MissingDependencyException,
     InvalidArgument,
     BentoMLException,
 )
-from bentoml.service_env import BentoServiceEnv
+from bentoml.service.env import BentoServiceEnv
 
 logger = logging.getLogger(__name__)
 
@@ -78,7 +64,7 @@ class OnnxModelArtifact(BentoServiceArtifact):
     >>>
     >>> import numpy
     >>> import bentoml
-    >>> from bentoml.artifact import OnnxModelArtifact
+    >>> from bentoml.frameworks.onnx import OnnxModelArtifact
     >>> from bentoml.adapters import DataframeInput
     >>>
     >>> @bentoml.env(auto_pip_dependencies=True)
