+++ bentoml/frameworks/pytorch.py
@@ -1,22 +1,8 @@
-# Copyright 2019 Atalaya Tech, Inc.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-# http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 import os
 import logging
 
-from bentoml.artifact import BentoServiceArtifact
-from bentoml.service_env import BentoServiceEnv
+from bentoml.service.artifacts import BentoServiceArtifact
+from bentoml.service.env import BentoServiceEnv
 from bentoml.utils import cloudpickle
 from bentoml.exceptions import MissingDependencyException, InvalidArgument
 
@@ -54,7 +40,7 @@ class PytorchModelArtifact(BentoServiceArtifact):
     >>>
     >>> import bentoml
     >>> from bentoml.adapters import ImageInput
-    >>> from bentoml.artifact import PytorchModelArtifact
+    >>> from bentoml.frameworks.pytorch import PytorchModelArtifact
     >>>
     >>> @bentoml.env(auto_pip_dependencies=True)
     >>> @bentoml.artifacts([PytorchModelArtifact('net')])
