+++ bentoml/frameworks/tensorflow.py
@@ -1,25 +1,11 @@
-# Copyright 2019 Atalaya Tech, Inc.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-# http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 import os
 import shutil
 import logging
 import pathlib
 
-from bentoml.artifact import BentoServiceArtifact
+from bentoml.service.artifacts import BentoServiceArtifact
 from bentoml.exceptions import MissingDependencyException
-from bentoml.service_env import BentoServiceEnv
+from bentoml.service.env import BentoServiceEnv
 
 logger = logging.getLogger(__name__)
 
@@ -174,7 +160,7 @@ class TensorflowSavedModelArtifact(BentoServiceArtifact):
     >>>
     >>> import bentoml
     >>> from bentoml.adapters import JsonInput
-    >>> from bentoml.artifact import TensorflowSavedModelArtifact
+    >>> from bentoml.frameworks.tensorflow import TensorflowSavedModelArtifact
     >>>
     >>> @bentoml.env(pip_dependencies=["tensorflow"])
     >>> @bentoml.artifacts([TensorflowSavedModelArtifact('model')])
