+++ bentoml/frameworks/xgboost.py
@@ -1,22 +1,8 @@
-# Copyright 2019 Atalaya Tech, Inc.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-# http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 import os
 
-from bentoml.artifact import BentoServiceArtifact
+from bentoml.service.artifacts import BentoServiceArtifact
 from bentoml.exceptions import MissingDependencyException, InvalidArgument
-from bentoml.service_env import BentoServiceEnv
+from bentoml.service.env import BentoServiceEnv
 
 
 class XgboostModelArtifact(BentoServiceArtifact):
@@ -43,7 +29,7 @@ class XgboostModelArtifact(BentoServiceArtifact):
     >>> model_to_save = xgboost.train(params=params, dtrain=dtrain)
     >>>
     >>> import bentoml
-    >>> from bentoml.artifact import XgboostModelArtifact
+    >>> from bentoml.frameworks.xgboost import XgboostModelArtifact
     >>> from bentoml.adapters import DataframeInput
     >>>
     >>> @bentoml.env(auto_pip_dependencies=True)
