+++ docs/source/quickstart.rst
@@ -80,7 +80,7 @@ the iris classifier model trained above:
 
     from bentoml import env, artifacts, api, BentoService
     from bentoml.adapters import DataframeInput
-    from bentoml.artifact import SklearnModelArtifact
+    from bentoml.frameworks.sklearn import SklearnModelArtifact
 
     @env(auto_pip_dependencies=True)
     @artifacts([SklearnModelArtifact('model')])
