+++ tests/bento_service_examples/onnx_onnxruntime_iris_classifier.py
@@ -1,7 +1,7 @@
 import numpy
 
 import bentoml
-from bentoml.artifact import OnnxModelArtifact
+from bentoml.frameworks.onnx import OnnxModelArtifact
 from bentoml.adapters import DataframeInput
